Complete center registration

+++ backend/src/main/java/com/example/isabloodbank/mapper/CenterMapper.java
@@ -1,13 +1,12 @@
 package com.example.isabloodbank.mapper;
 
 import com.example.isabloodbank.dto.CenterDTO;
-import com.example.isabloodbank.model.Address;
 import com.example.isabloodbank.model.Center;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 @Component
-public class CenterMapper implements ObjectMapper<Center, CenterDTO>{
+public class CenterMapper implements ObjectMapper<Center, CenterDTO> {
 
     @Autowired
     private AddressMapper addressMapper;
@@ -32,7 +31,7 @@ public class CenterMapper implements ObjectMapper<Center, CenterDTO>{
         center.setAddress(addressMapper.dtoToEntity(centerDTO.getAddress()));
         center.setDonationPrice(Long.parseLong(centerDTO.getDonationPrice()));
         center.setWorkingHours(centerDTO.getWorkingHours());
-        center.setRating(Long.parseLong(centerDTO.getRating()));
+        center.setRating(Double.parseDouble(centerDTO.getRating()));
         return center;
     }
 }
